refactor: move http/axios into api file
