+++ src/api/index.js
@@ -1,10 +1,53 @@
 import useSWRV from 'swrv'
+import axios from 'axios'
+import { get } from 'lodash'
+import localStorageCache from '@/composables/utils/localStorageCache'
+import alertStore from '@/composables/stores/alert'
+
+const $axios = axios.create({
+  baseURL: 'http://localhost:8080',
+  timeout: 3000,
+  crossDomain: true
+})
+const $auth = localStorageCache(0, 'app').get('auth')
+const user = $auth ? $auth.data : undefined
+if (user) { $axios.defaults.headers.common['Authorization'] = `BEARER ${user.token}` }
+
+/* provided methods */
+const http = (path, opts, handleErrors) => {
+  opts = opts || {}
+  const method = (opts.method || 'get').toLowerCase()
+  delete opts.method
+  const data = opts.data
+  delete opts.data
+
+  let req = (m => {
+    switch(m) {
+      case 'post':
+      case 'put':
+      case 'patch':
+        return $axios[method](path, data, opts)
+      default: return $axios[method](path, opts)
+    }
+  })(method)
+
+  const reqPromise = req.then(res => res.status === 200 ? res.data : res)
+
+  if (handleErrors) {
+    return reqPromise.catch(err => {
+      const msg = get(err, 'response.data.message')
+      if (msg) { alertStore.error(msg) }
+      return Promise.reject(err)
+    })
+  }
+  else { return reqPromise }
+}
 
 export const boardsApi = {
-  slugToBoardId: (http, slug) => {
+  slugToBoardId: (a, slug) => {
     return http(`/api/boards/${slug}/id`)
   },
-  getBoards: (http, { config, processBoardsCallback }) => {
+  getBoards: (a, { config, processBoardsCallback }) => {
     if (config) {
       let result = http('/api/boards')
       // use processor if available
@@ -23,43 +66,61 @@ export const boardsApi = {
 }
 
 export const threadsApi = {
-  byBoard: (http, opts) => {
+  byBoard: (a, opts) => {
     return http('/api/threads', opts)
   }
 }
 
 export const authApi = {
-  login: (http, opts, handleErrors) => { return http('/api/login', opts, handleErrors) },
-  logout: (http, opts, handleErrors) => { return http('/api/logout', opts, handleErrors) },
-  register: (http, opts, handleErrors) => { return http('/api/register', opts, handleErrors) },
-  emailAvailable: (http, val) => { return http(`/api/register/email/${val}`) },
-  usernameAvailable: (http, val) => { return http(`/api/register/username/${val}`) }
+  login: (a, opts, handleErrors) => {
+    return http('/api/login', opts, handleErrors)
+    .then(dbUser => {
+      $axios.defaults.headers.common['Authorization'] = `BEARER ${dbUser.token}`
+      return dbUser
+    })
+  },
+  logout: (a, opts, handleErrors) => {
+    return http('/api/logout', opts, handleErrors)
+    .then(dbUser => {
+      delete $axios.defaults.headers.common['Authorization']
+      return dbUser
+    })
+  },
+  register: (a, opts, handleErrors) => {
+    return http('/api/register', opts, handleErrors)
+    .then(dbUser => {
+      $axios.defaults.headers.common['Authorization'] = `BEARER ${dbUser.token}`
+      return dbUser
+    })
+  },
+  emailAvailable: (a, val) => { return http(`/api/register/email/${val}`) },
+  usernameAvailable: (a, val) => { return http(`/api/register/username/${val}`) }
 }
 
 export const usersApi = {
-  search: (http, username) => {
+  search: (a, username) => {
     return http(`/api/users/search?username=${username}`)
   },
-  update: (http, userId, opts) => {
+  update: (a, userId, opts) => {
     return http(`/api/users/${userId}`, opts)
   },
-  preferences: (http) => {
+  preferences: () => {
     return http('/api/users/preferences')
   }
 }
 
 export const breadcrumbsApi = {
-  find: (http, id, type) => {
+  find: (a, id, type) => {
     return http(`/api/breadcrumbs?id=${id}&type=${type}`)
   }
 }
 
 export const adminApi = {
   moderators: {
-    remove: (http, opts) => {
+    remove: (a, opts) => {
       return http('/api/admin/moderators/remove', opts)
     },
-    add: (http, opts) => {
+    add: (a, opts) => {
       return http('/api/admin/moderators', opts)
     }
   }
